requirements.txt
# 必须使用 Python 3.9 或更高版本！google-genai 库要求 Python 3.9+
PyQt6
aiohttp
openai
anthropic
google-genai  # 仅支持 Python 3.9 或更高版本，这是 Google 官方要求
configparser
requests
tqdm
qasync
networkx
matplotlib
# 知识库功能所需依赖
faiss-cpu  # 向量数据库
python-docx  # 处理DOCX文件
PyMuPDF  # 处理PDF文件，也称为fitz